exercise_4/requirements.txt
fastapi==0.111.1
numpy==1.24.4
opencv_python==4.9.0.80
Pillow==10.4.0
uvicorn
pytest